NEW/MVP/MCP/http-proxy/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Optional: for /render endpoint
# playwright>=1.40.0
